[WIP] More refactoring & bug fixes. + Forge updates (#109)
* Reformatting, inlining of variables, removed useless code artifacts..

* Removed useless axle_side model and used rotation instead.

* Override Block.getStateForPlacement().

That replaces the deprecated
Block.onBlockPlaced().
Also cached AABB of axles for improved performances.

* Override Block.getStateForPlacement().

That replaces the deprecated
Block.onBlockPlaced().
+cached AABB of axles for improved performances.
+updated mappings to 29/10/16

* BlockAxle now uses EnumFacing.Axis instead of arbitrary integers.

* Removed unused methods in interface.

* Refactored Windmills&Waterwheel to avoid code duplication.

* Refactored again parts of BlockAxle &Gens blocks.

* Reorganized proxy to more modern form and fixed typos.

* Removed direct references to optional classes, avoiding any undesired loading.

Also separated client methods for all compatibility modules.
They also have been slightly refactored to use an interface instead of an abstract class.

* Refactored Anchor and updated Forge version.

* Restored IMechanicalPower.getMinimalInput(EnumFacing).

* Added @SideOnly(Side.CLIENT) to all client methods in the same class as common methods.

+Updated to Forge 2121 and MCP Mappings 2016/11/05.

+++ src/main/java/betterwithmods/blocks/BlockCrank.java
@@ -62,7 +62,7 @@ public class BlockCrank extends BWMBlock implements IMechanicalBlock, IMultiVari
 
     @Override
     public boolean onBlockActivated(World world, BlockPos pos, IBlockState state, EntityPlayer player, EnumHand hand, ItemStack heldItem, EnumFacing side, float hitX, float hitY, float hitZ) {
-        int meta = state.getValue(STAGE).intValue();
+        int meta = state.getValue(STAGE);
 
         if (meta == 0) {
             if (player.getFoodStats().getFoodLevel() > 6) {
@@ -143,7 +143,7 @@ public class BlockCrank extends BWMBlock implements IMechanicalBlock, IMultiVari
 
     @Override
     public boolean isOutputtingMechPower(World world, BlockPos pos) {
-        return world.getBlockState(pos).getValue(STAGE).intValue() > 1;
+        return world.getBlockState(pos).getValue(STAGE) > 1;
     }
 
     @Override
@@ -173,7 +173,7 @@ public class BlockCrank extends BWMBlock implements IMechanicalBlock, IMultiVari
 
     @Override
     public void updateTick(World world, BlockPos pos, IBlockState state, Random rand) {
-        int stage = state.getValue(STAGE).intValue();
+        int stage = state.getValue(STAGE);
 
         if (stage > 0) {
             if (stage < 7) {
